+++ R/nfsa_get_data_all.R
@@ -6,7 +6,7 @@
 #' @param country Character vector specifying the country or countries to retrieve data for.
 #' @param table Character string specifying the table to retrieve data from (e.g., "T0801", "T0801SA", "T0800").
 #' @param filters one or several filters to be applied before collecting the data using all columns in the original parquet files.
-#' #' @param complete By default `FALSE` if we only want to bring `ref_area`,`id`,`time_period`,`obs_value`. If set to `TRUE` brings all columns.
+#' @param complete By default `FALSE` if we only want to bring `ref_area`,`id`,`time_period`,`obs_value`. If set to `TRUE` brings all columns.
 
 
 #' @return A tibble containing the selected NFSA data with columns: \code{ref_area}, \code{id}, \code{time_period}, and \code{obs_value}.
@@ -27,7 +27,8 @@
 nfsa_get_data_all <- function(input_sel = "M:/nas/Rprod/data/",
                               country,
                               table = "T0801",
-                              filters = NULL) {
+                              filters = NULL,
+                              complete = FALSE) {
 
   # 1. Setup Configuration -------------------------------------------------
   cfg_map <- list(
@@ -63,9 +64,7 @@ nfsa_get_data_all <- function(input_sel = "M:/nas/Rprod/data/",
   # 3. Data Processing Pipeline --------------------------------------------
   lookup <- nfsa::nfsa_sto_lookup
 
-  ds <- arrow::open_dataset(target_files) |>
-    dplyr::select(-embargo_date, -received)
-
+  ds <- arrow::open_dataset(target_files)
   # Apply Arrow filters (Push-down)
   if (!is.null(filters)) {
     parsed_filters <- rlang::parse_exprs(filters)
@@ -91,7 +90,6 @@ nfsa_get_data_all <- function(input_sel = "M:/nas/Rprod/data/",
       dplyr::left_join(lookup, by = c("counterpart_area", "ref_sector", "counterpart_sector",
                                       "consolidation", "accounting_entry", "sto",
                                       "instr_asset", "unit_measure", "prices")) |>
-      tidyr::drop_na() |>
       dplyr::arrange(time_period,ref_area,id) |>
       dplyr::mutate(ref_area = dplyr::if_else(ref_area == "GR", "EL", ref_area))
   }